Separate Compose UI from shared native core
