+++ app/build.gradle.kts
@@ -113,5 +113,5 @@ dependencies {
 
 
     implementation(projects.common)
+    implementation(projects.common.ui)
 }
-
