--- common/src/jvmMain/kotlin/dev/johnoreilly/commom/ui/ISSMapView.jvm.kt
@@ -1,9 +1,0 @@
-package dev.johnoreilly.common.ui
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import dev.johnoreilly.common.viewmodel.ISSPositionViewModel
-
-@Composable
-actual fun ISSMapView(modifier: Modifier, viewModel: ISSPositionViewModel) {
-}
