+++ common/src/mingwX64Main/kotlin/dev/johnoreilly/common/di/Koin.windows.kt
@@ -1,20 +1,19 @@
 package dev.johnoreilly.common.di
 
-import app.cash.sqldelight.driver.jdbc.sqlite.JdbcSqliteDriver
+import app.cash.sqldelight.async.coroutines.synchronous
+import app.cash.sqldelight.driver.native.NativeSqliteDriver
 import dev.johnoreilly.peopleinspace.db.PeopleInSpaceDatabase
 import io.ktor.client.engine.HttpClientEngine
-import io.ktor.client.engine.java.*
+import io.ktor.client.engine.winhttp.WinHttp
 import org.koin.dsl.module
 
 private class ContextWrapper
 
 actual fun nativeModule() = module {
-    includes(viewModelsModule())
     single { ContextWrapper() }
-    single<HttpClientEngine> { Java.create() }
+    single<HttpClientEngine> { WinHttp.create() }
     single {
-        val driver = JdbcSqliteDriver(JdbcSqliteDriver.IN_MEMORY)
-            .also { PeopleInSpaceDatabase.Schema.create(it) }
+        val driver = NativeSqliteDriver(PeopleInSpaceDatabase.Schema.synchronous(), "peopleinspace.db")
         PeopleInSpaceDatabaseWrapper(driver, PeopleInSpaceDatabase(driver))
     }
 }
