--- common/src/wasmJsMain/kotlin/com/surrus/common/ui/ISSMapView.wasmJs.kt
@@ -1,12 +1,0 @@
-package dev.johnoreilly.common.ui
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import dev.johnoreilly.common.viewmodel.ISSPositionViewModel
-
-@Composable
-actual fun ISSMapView(
-    modifier: Modifier,
-    viewModel: ISSPositionViewModel
-) {
-}
