+++ common/ui/build.gradle.kts
@@ -1,0 +1,86 @@
+@file:OptIn(ExperimentalWasmDsl::class)
+
+import org.jetbrains.kotlin.gradle.ExperimentalWasmDsl
+
+plugins {
+    alias(libs.plugins.kotlinMultiplatform)
+    alias(libs.plugins.android.kotlin.multiplatform.library)
+    alias(libs.plugins.jetbrainsCompose)
+    alias(libs.plugins.compose.compiler)
+    alias(libs.plugins.skie)
+    id("io.github.luca992.multiplatform-swiftpackage") version "2.3.0"
+}
+
+kotlin {
+    jvmToolchain(17)
+
+    listOf(
+        iosArm64(),
+        iosSimulatorArm64(),
+    ).forEach {
+        it.binaries.framework {
+            baseName = "common"
+            export(projects.common)
+        }
+    }
+
+    android {
+        namespace = "dev.johnoreilly.common.ui"
+        compileSdk = libs.versions.compileSdk.get().toInt()
+        minSdk = libs.versions.minSdk.get().toInt()
+        androidResources { enable = true }
+    }
+
+    jvm()
+
+    wasmJs {
+        browser()
+    }
+
+    sourceSets {
+        commonMain.dependencies {
+            api(projects.common)
+
+            implementation(compose.ui)
+            implementation(compose.runtime)
+            implementation(compose.foundation)
+            implementation(compose.material3)
+            implementation(compose.components.resources)
+            implementation(libs.androidx.lifecycle.compose.kmp)
+        }
+
+        commonTest.dependencies {
+            implementation(kotlin("test"))
+            @OptIn(org.jetbrains.compose.ExperimentalComposeLibrary::class)
+            implementation(compose.uiTest)
+        }
+
+        androidMain.dependencies {
+            implementation(libs.osmdroidAndroid)
+            implementation(libs.osm.android.compose)
+        }
+
+        jvmTest.dependencies {
+            implementation(compose.desktop.currentOs)
+        }
+    }
+}
+
+multiplatformSwiftPackage {
+    packageName("PeopleInSpaceKit")
+    swiftToolsVersion("5.9")
+    targetPlatforms {
+        iOS { v("14") }
+    }
+}
+
+kotlin.sourceSets.all {
+    languageSettings.optIn("kotlinx.cinterop.ExperimentalForeignApi")
+    languageSettings.optIn("kotlin.experimental.ExperimentalObjCName")
+}
+
+skie {
+    features {
+        enableSwiftUIObservingPreview = true
+    }
+}
