+++ common/ui/src/iOSMain/kotlin/com/surrus/common/ui/NativeViewFactory.kt
@@ -1,0 +1,8 @@
+package dev.johnoreilly.common.ui
+
+import dev.johnoreilly.common.viewmodel.ISSPositionViewModel
+import platform.UIKit.UIViewController
+
+interface NativeViewFactory {
+    fun createISSMapView(viewModel: ISSPositionViewModel): UIViewController
+}
