+++ settings.gradle.kts
@@ -26,5 +26,6 @@ include(":wearApp")
 include(":compose-desktop")
 include(":compose-web")
 include(":common")
+include(":common:ui")
 include(":backend")
 include(":mcp-server")
